requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0
Pillow==10.0.1
gunicorn